Add enums to testimonials
